% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openCVE.R
\docType{data}
\name{genes_WGCNA}
\alias{genes_WGCNA}
\title{Top 5000 most variant genes in TCGA RNAseq data}
\description{
A dataset containing the top 5000 most variant genes in TCGA RNAseq data for WGCNAmelanoma extension
}
\keyword{datasets}

