% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.oncotator.anno.R
\name{get.oncotator.anno}
\alias{get.oncotator.anno}
\title{Open Cancer Variant Explorer (CVE) Shiny app}
\usage{
get.oncotator.anno(x)
}
\arguments{
\item{x}{A matrix containing the columns chromosome, start, end, reference_allele and observed_allele.}
}
\description{
The get.oncotator.anno retrieves annotation from the Oncotator database.
}
\examples{
exCase <- data.frame(chr = rep(10, 3),
    start = c("100894110", "100985376", "101137905"),
    end = c("100894110", "100985376", "101137905"),
    ref_allele = c("T", "C", "G"),
    obs_allele = c("G", "A", "A"))
get.oncotator.anno(exCase)
}
