% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openCVE.R
\docType{data}
\name{GS_lscore}
\alias{GS_lscore}
\title{Lymphocyte score gene significance (GS)}
\description{
A dataset containing the lymphocyte score gene signficance for WGCNAmelanoma extension
}
\keyword{datasets}

