% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openCVE.R
\docType{data}
\name{GS_pmet}
\alias{GS_pmet}
\title{Primary vs metastasis gene significance (GS)}
\description{
A dataset containing the primary vs metastases gene signficance for WGCNAmelanoma extension
}
\keyword{datasets}

