% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openCVE.R
\docType{data}
\name{GS_Vem}
\alias{GS_Vem}
\title{Vemurafenib resistance gene significance (GS)}
\description{
A dataset containing the vemurafenib resistance gene signficance for WGCNAmelanoma extension
}
\keyword{datasets}

