% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openCVE.R
\docType{data}
\name{label_order}
\alias{label_order}
\title{Label order of co-expression modules}
\description{
A dataset containing the label order of co-expression modules for WGCNAmelanoma extension
}
\keyword{datasets}

