% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openCVE.R
\docType{data}
\name{METree_GO}
\alias{METree_GO}
\title{Gene tree of co-expression network}
\description{
A dataset containing the gene tree of co-expression network for WGCNAmelanoma extension
}
\keyword{datasets}

