% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openCVE.R
\docType{data}
\name{modules}
\alias{modules}
\title{Module assignment of top 5000 most variant genes in TCGA RNAseq data}
\description{
A dataset containing the module assignment top 5000 most variant genes in TCGA RNAseq data for WGCNAmelanoma extension
}
\keyword{datasets}

