% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openCVE.R
\docType{data}
\name{MS_lscore}
\alias{MS_lscore}
\title{Lymphocyte score module significance (MS)}
\description{
A dataset containing the lymphocyte score module signficance for WGCNAmelanoma extension
}
\keyword{datasets}

