% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openCVE.R
\docType{data}
\name{MS_lscore_bar}
\alias{MS_lscore_bar}
\title{Lymphocyte score module significance scaled for barplot}
\description{
A dataset containing the lymphocyte score module signficance scaled for barplot for WGCNAmelanoma extension
}
\keyword{datasets}

