% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openCVE.R
\docType{data}
\name{MS_pmet}
\alias{MS_pmet}
\title{Primary vs metastasis module significance (MS)}
\description{
A dataset containing the primary vs metastases module signficance for WGCNAmelanoma extension
}
\keyword{datasets}

