% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openCVE.R
\docType{data}
\name{MS_pmet_bar}
\alias{MS_pmet_bar}
\title{Primary vs metastasis module significance scaled for barplot}
\description{
A dataset containing the primary vs metastases module signficance scaled for barplot for WGCNAmelanoma extension
}
\keyword{datasets}

