% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openCVE.R
\docType{data}
\name{MS_survival}
\alias{MS_survival}
\title{Survival module significance (MS)}
\description{
A dataset containing the survival module signficance for WGCNAmelanoma extension
}
\keyword{datasets}

