% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openCVE.R
\docType{data}
\name{MS_survival_bar}
\alias{MS_survival_bar}
\title{Survival module significance scaled for barplot}
\description{
A dataset containing the survival module signficance scaled for barplot for WGCNAmelanoma extension
}
\keyword{datasets}

