% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openCVE.R
\docType{data}
\name{MS_UV_bar}
\alias{MS_UV_bar}
\title{UV signature module significance scaled for barplot}
\description{
A dataset containing the UV signature module signficance scaled for barplot for WGCNAmelanoma extension
}
\keyword{datasets}

