% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openCVE.R
\docType{data}
\name{MS_vem}
\alias{MS_vem}
\title{Vemurafenib resistance module significance (MS)}
\description{
A dataset containing the vemurafenib resistance module signficance for WGCNAmelanoma extension
}
\keyword{datasets}

