% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openCVE.R
\docType{data}
\name{MS_Vem_bar}
\alias{MS_Vem_bar}
\title{Vemurafenib resistance module significance scaled for barplot}
\description{
A dataset containing the vemurafenib resistance module signficance scaled for barplot for WGCNAmelanoma extension
}
\keyword{datasets}

