% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openCVE.R
\docType{data}
\name{oncotator_example}
\alias{oncotator_example}
\title{Example Oncotator output for the case study described in the paper}
\description{
A dataset containing the example Oncotator output for the case study described in the paper
}
\keyword{datasets}

