% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openCVE.R
\name{openCVE}
\alias{openCVE}
\title{Open Cancer Variant Explorer (CVE) Shiny app}
\usage{
openCVE(x, sample_names, extension = FALSE)
}
\arguments{
\item{x}{A dataframe (for single file) or list (for multiple oncotator output files)}

\item{sample_names}{A character vector with sample name(s)}

\item{extension}{A character vector of extention name}
}
\value{opens CVE Shiny app}
\description{
The openCVE function opens the CVE Shiny app.
The function to supplement the R package with the Shiny app was suggested by Dean Attali (http://deanattali.com). Currently, the only extension available is a melanoma co-expression network (WGCNAmelanoma).
}
\examples{
\donttest{
openCVE(oncotator_example,"case study")
openCVE(oncotator_example,"case study WGCNA","WGCNAmelanoma")
}
}

