% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openCVE.R
\docType{data}
\name{WGCNAmelanoma_extension}
\alias{WGCNAmelanoma_extension}
\title{Melanoma co-expression network CVE extension}
\description{
A dataset containing the files to explore variant genes in melanoma-specific co-expression networks within CVE. A proof of principle how the files were generated is described in the vignette WGCNA_from_TCGA_RNAseq.
In summary the files are:

Gene significance measures: GS_lscore, GS_pmet, GS_survival, GS_UV, GS_Vem

Module membership: MM

Module significance measures: MS_lscore_bar, MS_pmet_bar, MS_survival_bar, MS_UV_bar, MS_Vem_bar

Top 5000 mad genes of WGCNA: genes_WGCNA

label order of modules: label_order

gene tree of co-expression network: METree_GO

module assignment of genes: modules

}
\keyword{datasets}

