% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.oncotator.anno.R
\name{get.oncotator.anno}
\alias{get.oncotator.anno}
\title{Open Cancer Variant Explorer (CVE) Shiny app}
\usage{
get.oncotator.anno(x)
}
\arguments{
\item{x}{A matrix containing the columns chromosome, start, end, reference_allele and observed_allele.}
}
\description{
The get.oncotator.anno retrieves annotation from the Oncotator database.
}
\examples{
get.oncotator.anno(crcCase_input)
}

